restructures redux-related code to be in concatenated dir (see #13) breaking changes start here, bumped version up
