+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-redux-app",
-  "version": "0.4.0",
+  "version": "1.0.0",
   "description": "An opinionated development-ready React-Redux boilerplate that will instantly get you going.",
   "main": "index.html",
   "scripts": {
